+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Changed
   * Updated role to be `undefined` for empty cell.
 
+* Changed
+  * Removed terra-props-table dev dependency.
+
 ## 2.28.0 - (August 11, 2020)
 
 * Added