+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Changed
-  * Updated role to be `undefined` for empty cell.
+  * Updated role to be `undefined` for non interactable cells.
 
 * Changed
   * Removed terra-props-table dev dependency.