+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Changed
+  * Updated role to be `undefined` for non interactable cells.
+
+* Changed
   * Removed terra-props-table dev dependency.
 
 ## 2.28.0 - (August 11, 2020)